refactor(pm): restore cmd-layer boundaries for clean, view, list
Per the layering rules (cmd/ assembles and delegates; business logic
in service/; display in util/format_print.rs): cache cleaning moves to
service/clean_cache.rs with a reusable confirm() prompt helper;
view's print_*/format_* helpers move to format_print (kept on
owo_colors in a private submodule to preserve byte-exact output);
list's tree building joins LockGraphService and its renderer joins
format_print, swapping raw ANSI escapes for the colored crate.
